tests: move tests ran on GitHub to end-to-end workflow
